libcontainer: move capabilities to separate package
